🌱 Port the viewer's store reads to Rust
`view/store.py` is the seam every page reads through, and it carries four obligations the port
had to keep whole: a read-only connection per request, the macro library installed on it, the
schema version checked before any query, and a store under someone else's write lock answered
as a 503 rather than a crash. `hyphae-view/src/store.rs` reproduces all four.

The catalog ports as three enums rather than one, because which of them a query is declared in
is what says whether it may return a fat column whole — the split is a rule, not filing.
Keyset paging and the `_core`/window composition come with it: the window counts its pre-LIMIT
matches into `matched_rows`, a name of the composition rather than of any library query, so the
query a page cites stays the unlimited one on disk.

`Param` joins `Row` in `hyphae-store` as the binding half of that seam. It is the closed set
`ParamValue` declares in Python rather than a `&dyn ToSql`, so a call site reads as the binding
it is; it also keeps `duckdb` out of `hyphae-view`'s dependencies entirely.

One trap found by serving a real session: DuckDB answers a `SUM()` over a BIGINT column with a
HUGEINT, so every rolled-up count arrived as a type `Row::i64` refused. Narrowed rather than
carried, since Python reads the same value as a plain `int`.

+++ rust/crates/hyphae-store/src/queries.rs
@@ -29,29 +29,29 @@ pub const VIEW_PREFIX: &str = "view_";
 /// How much of a title a NavTree row carries — a turn's, a run's, an api call's. What a row
 /// *can* say rather than what fits: the NavTree is draggable, so the cut has to survive a
 /// reader widening it.
-pub const NAV_CHARS: i64 = 110;
+pub const NAV_CHARS: usize = 110;
 
 /// How much of a title one crumb of a crumb chain carries. A crumb is a place to click, not a
 /// place to read — the node itself is open underneath.
-pub const CRUMB_CHARS: i64 = 40;
+pub const CRUMB_CHARS: usize = 40;
 
 /// How much of a string a node page's header carries, and how many items and how much of each
 /// one a header list holds.
-pub const HEADER_CHARS: i64 = 100;
-pub const HEADER_ITEMS: i64 = 5;
-pub const HEADER_ITEM_CHARS: i64 = 60;
+pub const HEADER_CHARS: usize = 100;
+pub const HEADER_ITEMS: usize = 5;
+pub const HEADER_ITEM_CHARS: usize = 60;
 
 /// How much of a string one row of a children log carries.
-pub const LOG_CHARS: i64 = 300;
+pub const LOG_CHARS: usize = 300;
 
 /// How many rows one children log holds.
 pub const LOG_ROWS: i64 = 100;
 
 /// How much of the one value a node page is *about* the pane shows before it offers the rest.
-pub const DETAIL_CHARS: i64 = 4_000;
+pub const DETAIL_CHARS: usize = 4_000;
 
 /// How much of a model-written description or friction line a page shows.
-pub const ENRICHMENT_CHARS: i64 = 200;
+pub const ENRICHMENT_CHARS: usize = 200;
 
 /// The keyset cursor before the first row: "the last index already shown", and indexes start
 /// at 0.
